backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.9
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
pypdf>=5.0.0
python-docx>=1.0.0
